only try to copy the binary to make the LED blink if it exists.

+++ copy_to_final.sh
@@ -22,7 +22,9 @@ fi
 tar -cvpzf backup.tar.gz -C fs/ .
 tar -xzf backup.tar.gz -C ${CHROOT_PATH}/
 chown root:root ${CHROOT_PATH}/etc/sudoers
-cp deb/led ${CHROOT_PATH}/usr/sbin/led
+if [ -f deb/led ]; then
+    cp deb/led ${CHROOT_PATH}/usr/sbin/led
+fi
 cp -rf midi ${CHROOT_PATH}/opt/
 
 # needed just once, make sure the size of the partition is correct
